refactor: bind cursor queries to captured syntax domains
Preserve captured document and engine identity through hover and navigation, isolate unresolved syntax domains, and measure production request stages. Fix guarded header hover selection and canceled parser reuse found by real replay. P3 reviewer gate passed; 384 binding requests and U-Boot memory/completion gates pass. Full index: U-Boot 56.370 s, Wine 110.625 s; completion P95 29.772 ms; warm publication peak 507645952 bytes. Final version-wide verification remains required after remaining proposals.

+++ crates/fossilsense/src/store_parser_adapter.rs
@@ -16,7 +16,8 @@ impl PersistableFileIndex for FileSemanticIndex {
 
     fn persistence_diagnostics(&self) -> PersistenceDiagnostics {
         PersistenceDiagnostics {
-            fact_mask: self.diagnostics.requested_facts.bits(),
+            // Only the original durable fact groups cross the storage boundary.
+            fact_mask: self.diagnostics.requested_facts.bits() as u8,
             parse_error_count: self.diagnostics.parse_error_count,
             fallback_used: self.diagnostics.fallback_used,
         }
